requirements.txt
python-dotenv
requests
beautifulsoup4
fastapi
uvicorn
